Let IpHistoryRepository use the database class
- Add tests for the repo
- Extract user-ip-history creation from User into a separate class
- Some type juggling

+++ src/Plugin/AmpacheRSSView.php
@@ -110,7 +110,7 @@ class AmpacheRSSView
                 echo '<tr><td>';
                 echo '<div>';
                 echo '<div style="float: left; font-weight: bold;"><a href="' . $item->link . '" target="_blank">' . $item->title . '</a></div>';
-                echo '<div style="float: right;">' . get_datetime(strtotime($item->pubDate), 'short', 'short', "m/d/Y H:i") . '</div>';
+                echo '<div style="float: right;">' . get_datetime((int) strtotime($item->pubDate), 'short', 'short', "m/d/Y H:i") . '</div>';
                 echo '</div><br />';
                 echo '<div style="margin-left: 30px;">';
                 if (isset($item->image)) {
